pyproject.toml
[project]
name = "new-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "django>=6.0",
    "django-crispy-forms>=2.5",
    "django-froala-editor>=3.0",
    "Pillow>=10.0",
]
